+++ api/src/main/java/org/accula/api/model/GitPullRequest.java
@@ -1,28 +1,10 @@
 package org.accula.api.model;
 
+import lombok.Data;
+
+@Data
 public class GitPullRequest {
     private Integer number;
     private String url;
     private GitUserModel user;
-
-    public GitPullRequest(){
-        super();
-    }
-
-    public void setUser(final GitUserModel user){
-        this.user = user;
-    }
-
-    public GitUserModel getUser() {
-        return user;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public Integer getNumber(){
-        return number;
-    }
-
 }
