+++ api/src/main/java/org/accula/api/routers/WebHookRouter.java
@@ -1,7 +1,7 @@
 package org.accula.api.routers;
 
 import lombok.RequiredArgsConstructor;
-import org.accula.api.handlers.DataParserHandler;
+import org.accula.api.handlers.WebHookHandler;
 import org.springframework.context.annotation.Bean;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.server.RouterFunction;
@@ -10,8 +10,8 @@ import org.springframework.web.reactive.function.server.ServerResponse;
 
 @Component
 @RequiredArgsConstructor
-public class DataParserRouter {
-    private final DataParserHandler handler;
+public class WebHookRouter {
+    private final WebHookHandler handler;
 
     @Bean
     public RouterFunction<ServerResponse> dataParserRoute() {
